+++ src/app/modules/+home/home-routing.module.ts
@@ -1,11 +1,12 @@
 import { NgModule } from '@angular/core';
 import { Routes, RouterModule } from '@angular/router';
 import { HomeContainerComponent } from './containers/home-container/home-container.component';
+import { STRING_EMPTY } from '@shared/constants/string.constants';
 
 
 const routes: Routes = [
   {
-    path: '',
+    path: STRING_EMPTY,
     component: HomeContainerComponent
   }
 ];