+++ src/app/shared/constants/enviroment.constants.ts
@@ -1,0 +1,2 @@
+export const BACK_LOCAL_URL = 'http://localhost:5208';
+export const BACK_PROD_URL = 'https://www.mybackend.com';