+++ src/app/shared/models/language.models.ts
@@ -1,0 +1,4 @@
+export enum LanguageEnum {
+  SPANISH = 'es',
+  ENGLISH = 'en',
+}