+++ CHANGELOG.md
@@ -18,3 +18,16 @@
   - Footer Component added.
   - Logger Component added.
   - App colors added.
+
+
+## TODO IN BACK:
+  - Relate users with google users (Maybe not in front)
+  - Add surname2 to user? Not sure about that
+  - delete password from user
+  - GET prices by article
+  - GET user by GoogleId or Login using google token
+  - add languaje to user
+
+## TODO IN FRONT:
+  - minimize weight of images
+  - Social Links